audio-producer/requirements.txt
flask>=3.0.0
numpy>=1.26.0
scipy>=1.11.4
pydub>=0.25.1

# Phase 9: Additional dependencies for genre diversity
mido>=1.2.10
soundfile>=0.12.1
pyyaml>=6.0
requests>=2.31.0
